requirements.txt
python-telegram-bot
gspread
google-auth
apscheduler
python-dotenv